examlyzer-backend/app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
python-dotenv
jose
bcrypt